src/requirements.txt
dvclive>=3.0
pandas
pyaml
scikit-learn>=1.3
scipy
matplotlib
beautifulsoup4==4.12.3
